Use common Server and DataModelHandler from common/chip_app_server Move Led_Widget to platform/EFR Add efr lighting app to the CI test
