+++ .github/workflows/examples.yaml
@@ -143,6 +143,8 @@ jobs:
                   submodules: true
             - name: Build example EFR32 Lock App for BRD4161A
               run: scripts/examples/gn_efr32_example.sh examples/lock-app/efr32/ out/lock_app_debug BRD4161A
+            - name: Build example EFR32 Lighting App for BRD4161A
+              run: scripts/examples/gn_efr32_example.sh examples/lighting-app/efr32/ out/lighting_app_debug BRD4161A
             - name: Binary artifact suffix
               id: outsuffix
               uses: haya14busa/action-cond@v1.0.0
@@ -158,6 +160,7 @@ jobs:
                       steps.outsuffix.outputs.value }}
                   path: |
                       out/lock_app_debug/BRD4161A/chip-efr32-lock-example.out
+                      out/lighting_app_debug/BRD4161A/chip-efr32-lighting-example.out
 
     nrfconnect:
         name: nRF Connect SDK
