+++ examples/lighting-app/efr32/BUILD.gn
@@ -64,6 +64,7 @@ efr32_executable("lighting_app") {
     ":sdk",
     "${chip_root}/src/lib",
     "${chip_root}/src/setup_payload",
+    "${chip_root}/examples/common/chip-app-server:chip-app-server",
     "${chip_root}/third_party/openthread/platforms:libopenthread-platform",
     "${chip_root}/third_party/openthread/platforms:libopenthread-platform-utils",
     "${chip_root}/third_party/openthread/repo:libopenthread-cli-ftd",
@@ -79,6 +80,7 @@ efr32_executable("lighting_app") {
 
   sources = [
     "${chip_root}/examples/common/QRCode/repo/c/qrcodegen.c",
+    "${chip_root}/examples/common/chip-app-server/DataModelHandler.cpp",
     "${chip_root}/src/app/clusters/on-off-server/on-off.c",
     "${chip_root}/src/app/util/af-event.cpp",
     "${chip_root}/src/app/util/attribute-size.c",
@@ -96,12 +98,12 @@ efr32_executable("lighting_app") {
     "${examples_plat_dir}/${efr32_family}/${efr32_board}/init_board.c",
     "${examples_plat_dir}/${efr32_family}/${efr32_board}/init_mcu.c",
     "${examples_plat_dir}/init_otSystem.c",
+    "${examples_plat_dir}/LEDWidget.cpp",
+    "${examples_plat_dir}/Service.cpp",
     "src/AppTask.cpp",
     "src/ButtonHandler.cpp",
-    "src/DataModelHandler.cpp",
-    "src/LEDWidget.cpp",
+    "src/ZclCallbacks.cpp",
     "src/LightingManager.cpp",
-    "src/Server.cpp",
     "src/gen/call-command-handler.c",
     "src/gen/callback-stub.c",
     "src/gen/znet-bookkeeping.c",
