+++ examples/lighting-app/efr32/README.md
@@ -1,4 +1,4 @@
-# CHIP EFR32 Lock Example
+# CHIP EFR32 Lighting Example
 
 An example showing the use of CHIP on the Silicon Labs EFR32 MG12.
 
@@ -106,7 +106,7 @@ implemented inside the lighting-app example.
 
     -   From CHIP root,
 
-              $ python out/lock_app_debug/BRD4161A/chip-efr32-lighting-example.flash.py
+              $ python out/lighting_app_debug/BRD4161A/chip-efr32-lighting-example.flash.py
 
 -   Or with the Ozone debugger, just load the .out file.
 
@@ -172,7 +172,7 @@ combination with JLinkRTTClient as follows:
 -   Once said connectection is established (you can verify that with the command
     `router table` using a serial terminal (screen / minicom etc.) on the board
     running the lighting-app example)
--   Using chip-tool you can now control the lock status with on/off command such
+-   Using chip-tool you can now control the light status with on/off command such
     as `chip-tool on <ipv6 address of the node> 11095 1`
 
 ### Notes
