+++ examples/lighting-app/efr32/include/AppTask.h
@@ -60,10 +60,6 @@ private:
     static void LightActionEventHandler(AppEvent * aEvent);
     static void TimerEventHandler(TimerHandle_t xTimer);
 
-    static void HandleBLEConnectionOpened(chip::Ble::BLEEndPoint * endPoint);
-    static void HandleBLEConnectionClosed(chip::Ble::BLEEndPoint * endPoint, BLE_ERROR err);
-    static void HandleBLEMessageReceived(chip::Ble::BLEEndPoint * endPoint, chip::System::PacketBuffer * buffer);
-
     static void UpdateClusterState(void);
 
     void StartTimer(uint32_t aTimeoutMs);
