+++ examples/lighting-app/efr32/src/AppTask.cpp
@@ -24,6 +24,7 @@
 #include "DataModelHandler.h"
 #include "LEDWidget.h"
 #include "Server.h"
+#include "Service.h"
 #include "attribute-storage.h"
 #include "gen/cluster-id.h"
 #include "lcd.h"
@@ -94,6 +95,9 @@ int AppTask::Init()
 {
     CHIP_ERROR err = CHIP_NO_ERROR;
 
+    // Init ZCL Data Model
+    InitServer();
+
     // Initialise WSTK buttons PB0 and PB1 (including debounce).
     ButtonHandler::Init();
 
@@ -133,6 +137,8 @@ int AppTask::Init()
     chip::SetupPayload payload;
     uint32_t setUpPINCode       = 0;
     uint16_t setUpDiscriminator = 0;
+    uint16_t vendorId           = 0;
+    uint16_t productId          = 0;
 
     err = ConfigurationMgr().GetSetupPinCode(setUpPINCode);
     if (err != CHIP_NO_ERROR)
@@ -146,9 +152,21 @@ int AppTask::Init()
         EFR32_LOG("ConfigurationMgr().GetSetupDiscriminator() failed: %s", chip::ErrorStr(err));
     }
 
+    err = ConfigurationMgr().GetVendorId(vendorId);
+    if (err != CHIP_NO_ERROR)
+    {
+        EFR32_LOG("ConfigurationMgr().GetVendorId() failed: %s", chip::ErrorStr(err));
+    }
+
+    err = ConfigurationMgr().GetProductId(productId);
+    if (err != CHIP_NO_ERROR)
+    {
+        EFR32_LOG("ConfigurationMgr().GetProductId() failed: %s", chip::ErrorStr(err));
+    }
+
     payload.version       = 1;
-    payload.vendorID      = EXAMPLE_VENDOR_ID;
-    payload.productID     = 1;
+    payload.vendorID      = vendorId;
+    payload.productID     = productId;
     payload.setUpPINCode  = setUpPINCode;
     payload.discriminator = setUpDiscriminator;
     chip::QRCodeSetupPayloadGenerator generator(payload);
@@ -167,72 +185,6 @@ int AppTask::Init()
     return err;
 }
 
-void AppTask::HandleBLEConnectionOpened(chip::Ble::BLEEndPoint * endPoint)
-{
-    assert(endPoint != NULL);
-
-    ChipLogProgress(DeviceLayer, "AppTask: Connection opened");
-
-    GetAppTask().mBLEEndPoint    = endPoint;
-    endPoint->OnMessageReceived  = AppTask::HandleBLEMessageReceived;
-    endPoint->OnConnectionClosed = AppTask::HandleBLEConnectionClosed;
-}
-
-void AppTask::HandleBLEConnectionClosed(chip::Ble::BLEEndPoint * endPoint, BLE_ERROR err)
-{
-    ChipLogProgress(DeviceLayer, "AppTask: Connection closed");
-
-    GetAppTask().mBLEEndPoint = nullptr;
-}
-
-void AppTask::HandleBLEMessageReceived(chip::Ble::BLEEndPoint * endPoint, chip::System::PacketBuffer * buffer)
-{
-    assert(endPoint != NULL);
-#if CHIP_ENABLE_OPENTHREAD
-    assert(buffer != NULL);
-
-    uint16_t bufferLen = buffer->DataLength();
-    uint8_t * data     = buffer->Start();
-    chip::DeviceLayer::Internal::DeviceNetworkInfo networkInfo;
-    ChipLogProgress(DeviceLayer, "AppTask: Receive message size %u", bufferLen);
-
-    memcpy(networkInfo.ThreadNetworkName, data, sizeof(networkInfo.ThreadNetworkName));
-    data += sizeof(networkInfo.ThreadNetworkName);
-
-    memcpy(networkInfo.ThreadExtendedPANId, data, sizeof(networkInfo.ThreadExtendedPANId));
-    data += sizeof(networkInfo.ThreadExtendedPANId);
-
-    memcpy(networkInfo.ThreadMeshPrefix, data, sizeof(networkInfo.ThreadMeshPrefix));
-    data += sizeof(networkInfo.ThreadMeshPrefix);
-
-    memcpy(networkInfo.ThreadNetworkKey, data, sizeof(networkInfo.ThreadNetworkKey));
-    data += sizeof(networkInfo.ThreadNetworkKey);
-
-    memcpy(networkInfo.ThreadPSKc, data, sizeof(networkInfo.ThreadPSKc));
-    data += sizeof(networkInfo.ThreadPSKc);
-
-    networkInfo.ThreadPANId = data[0] | (data[1] << 8);
-    data += sizeof(networkInfo.ThreadPANId);
-    networkInfo.ThreadChannel = data[0];
-    data += sizeof(networkInfo.ThreadChannel);
-
-    networkInfo.FieldPresent.ThreadExtendedPANId = *data;
-    data++;
-    networkInfo.FieldPresent.ThreadMeshPrefix = *data;
-    data++;
-    networkInfo.FieldPresent.ThreadPSKc = *data;
-    data++;
-    networkInfo.NetworkId              = 0;
-    networkInfo.FieldPresent.NetworkId = true;
-
-    ThreadStackMgr().SetThreadEnabled(false);
-    ThreadStackMgr().SetThreadProvision(networkInfo);
-    ThreadStackMgr().SetThreadEnabled(true);
-#endif
-    endPoint->Close();
-    chip::System::PacketBuffer::Free(buffer);
-}
-
 void AppTask::AppTaskMain(void * pvParameter)
 {
     int err;
@@ -247,7 +199,6 @@ void AppTask::AppTaskMain(void * pvParameter)
     }
 
     EFR32_LOG("App Task started");
-    chip::DeviceLayer::ConnectivityMgr().AddCHIPoBLEConnectionHandler(&AppTask::HandleBLEConnectionOpened);
     SetDeviceName("EFR32LightingDemo._chip._udp.local.");
 
     while (true)
