--- examples/lock-app/efr32/src/LEDWidget.cpp
@@ -1,95 +1,0 @@
-/*
- *
- *    Copyright (c) 2020 Project CHIP Authors
- *    Copyright (c) 2019 Google LLC.
- *    All rights reserved.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#include "LEDWidget.h"
-#include "AppTask.h"
-
-#include "bsp.h"
-
-#include <platform/CHIPDeviceLayer.h>
-
-void LEDWidget::InitGpio(void)
-{
-    // Sets gpio pin mode for ALL board Leds.
-    BSP_LedsInit();
-}
-
-void LEDWidget::Init(int ledNum)
-{
-    mLastChangeTimeUS = 0;
-    mBlinkOnTimeMS    = 0;
-    mBlinkOffTimeMS   = 0;
-    mLedNum           = ledNum;
-
-    Set(false);
-}
-
-void LEDWidget::Invert(void)
-{
-    Set(!mState);
-}
-
-void LEDWidget::Set(bool state)
-{
-    mBlinkOnTimeMS  = 0;
-    mBlinkOffTimeMS = 0;
-    DoSet(state);
-}
-
-void LEDWidget::Blink(uint32_t changeRateMS)
-{
-    Blink(changeRateMS, changeRateMS);
-}
-
-void LEDWidget::Blink(uint32_t onTimeMS, uint32_t offTimeMS)
-{
-    mBlinkOnTimeMS  = onTimeMS;
-    mBlinkOffTimeMS = offTimeMS;
-    Animate();
-}
-
-void LEDWidget::Animate()
-{
-    if (mBlinkOnTimeMS != 0 && mBlinkOffTimeMS != 0)
-    {
-        int64_t nowUS            = ::chip::System::Layer::GetClock_MonotonicHiRes();
-        int64_t stateDurUS       = ((mState) ? mBlinkOnTimeMS : mBlinkOffTimeMS) * 1000LL;
-        int64_t nextChangeTimeUS = mLastChangeTimeUS + stateDurUS;
-
-        if (nowUS > nextChangeTimeUS)
-        {
-            DoSet(!mState);
-            mLastChangeTimeUS = nowUS;
-        }
-    }
-}
-
-void LEDWidget::DoSet(bool state)
-{
-    mState = state;
-
-    if (state)
-    {
-        BSP_LedSet(mLedNum);
-    }
-    else
-    {
-        BSP_LedClear(mLedNum);
-    }
-}
