+++ examples/lock-app/efr32/src/ZclCallbacks.cpp
@@ -20,17 +20,17 @@
  *   This file implements the handler for data model messages.
  */
 
-#include <system/SystemPacketBuffer.h>
-
 #include "AppConfig.h"
-#include "DataModelHandler.h"
-#include "LightingManager.h"
+#include "BoltLockManager.h"
 
-#include "af-types.h"
 #include "gen/attribute-id.h"
 #include "gen/cluster-id.h"
+#include <app/chip-zcl-zpro-codec.h>
+#include <app/util/af-types.h>
+#include <app/util/attribute-storage.h>
+#include <app/util/util.h>
 
-using namespace ::chip;
+// using namespace ::chip;
 
 extern "C" void emberAfPostAttributeChangeCallback(uint8_t endpoint, EmberAfClusterId clusterId, EmberAfAttributeId attributeId,
                                                    uint8_t mask, uint16_t manufacturerCode, uint8_t type, uint8_t size,
@@ -50,10 +50,10 @@ extern "C" void emberAfPostAttributeChangeCallback(uint8_t endpoint, EmberAfClus
 
     if (*value)
     {
-        LightMgr().InitiateAction(AppEvent::kEventType_Light, LightingManager::ON_ACTION);
+        BoltLockMgr().InitiateAction(AppEvent::kEventType_Lock, BoltLockManager::LOCK_ACTION);
     }
     else
     {
-        LightMgr().InitiateAction(AppEvent::kEventType_Light, LightingManager::OFF_ACTION);
+        BoltLockMgr().InitiateAction(AppEvent::kEventType_Lock, BoltLockManager::UNLOCK_ACTION);
     }
 }
