+++ examples/platform/efr32/LEDWidget.h
